pyproject.toml
[tool.poetry]
name = "quotes_service"
version = "0.1.0"
description = ""
authors = ["Elisei"]

[tool.poetry.dependencies]
python = "^3.11"
pydantic = "^2.3.0"
fastapi = "^0.103.0"
orjson = "^3.9.2"
certifi = "^2023.7.22"
uvicorn = "^0.23.2"
python-dotenv = "^1.0.0"
pydantic-settings = "^2.0.2"

[tool.poetry.group.dev.dependencies]
httpx = "^0.25.0"
pytest = "^7.4.0"
pytest-asyncio = "^0.21.1"
mypy = "^1.4.1"
isort = "^5.12.0"
flake8 = "^6.1.0"
black = "^23.7.0"
ruff = "^0.0.291"

[tool.mypy]
strict = true
ignore_missing_imports = true

[tool.black]
line-length = 120
target_version = ['py311']

[tool.isort]
profile = "django"
lines_after_imports = 2
line_length = 120
src_paths = ["app"]

[tool.ruff]
line-length = 120

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
